Updating BPS BAM Publisher and related configurations to be compatible with DAS. Issue - https://wso2.org/jira/browse/BPS-698
