+++ components/bpel/org.wso2.carbon.bpel.analytics.publisher/src/main/java/org/wso2/carbon/bpel/analytics/publisher/AnalyticsPublisherExtensionBundle.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) , WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.wso2.carbon.bpel.analytics.publisher;
+
+import org.apache.ode.bpel.runtime.extension.AbstractExtensionBundle;
+
+/**
+ * Class that bundles and  registers <code>&lt;BPActivity&gt;</code> to Analytics server based activity monitoring on a bpel
+ * process deployed BPS
+ */
+
+
+public class AnalyticsPublisherExtensionBundle extends AbstractExtensionBundle {
+
+    @Override
+    public String getNamespaceURI() {
+        return AnalyticsPublisherConstants.ANALYTICS_PUBLISHER_NS;
+    }
+
+    @Override
+    public void registerExtensionActivities() {
+        registerExtensionOperation("publish", AnalyticsPublisherExtensionOperation.class);
+    }
+}
