+++ components/bpel/org.wso2.carbon.bpel.analytics.publisher/src/main/java/org/wso2/carbon/bpel/analytics/publisher/internal/BPELAnalyticsPublisherContentHolder.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright (c) 2012, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.wso2.carbon.bpel.analytics.publisher.internal;
+
+import org.wso2.carbon.bpel.core.ode.integration.BPELServer;
+import org.wso2.carbon.registry.core.service.RegistryService;
+
+public class BPELAnalyticsPublisherContentHolder {
+
+    private static BPELAnalyticsPublisherContentHolder instance;
+
+    private BPELServer bpelServer;
+    private RegistryService registryService;
+
+    private BPELAnalyticsPublisherContentHolder(){}
+
+    public static BPELAnalyticsPublisherContentHolder getInstance(){
+        if(null == instance){
+            instance = new BPELAnalyticsPublisherContentHolder();
+        }
+        return instance;
+    }
+
+    public BPELServer getBpelServer(){
+        return bpelServer;
+    }
+
+    public void setBpelServer(BPELServer bpelServer){
+        this.bpelServer = bpelServer;
+    }
+
+    public  void setRegistryService(RegistryService registryService){
+        this.registryService = registryService;
+    }
+
+    public RegistryService getRegistryService() {
+        return registryService;
+    }
+}
