--- components/bpel/org.wso2.carbon.bpel.bam.publisher/src/main/java/org/wso2/carbon/bpel/bam/publisher/internal/BPELBamPublisherContentHolder.java
@@ -1,53 +1,0 @@
-/*
- * Copyright (c) 2012, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.wso2.carbon.bpel.bam.publisher.internal;
-
-import org.wso2.carbon.bpel.core.ode.integration.BPELServer;
-import org.wso2.carbon.registry.core.service.RegistryService;
-
-public class BPELBamPublisherContentHolder {
-
-    private static BPELBamPublisherContentHolder instance;
-
-    private BPELServer bpelServer;
-    private RegistryService registryService;
-
-    private BPELBamPublisherContentHolder(){}
-
-    public static BPELBamPublisherContentHolder getInstance(){
-        if(null == instance){
-            instance = new BPELBamPublisherContentHolder();
-        }
-        return instance;
-    }
-
-    public BPELServer getBpelServer(){
-        return bpelServer;
-    }
-
-    public void setBpelServer(BPELServer bpelServer){
-        this.bpelServer = bpelServer;
-    }
-
-    public  void setRegistryService(RegistryService registryService){
-        this.registryService = registryService;
-    }
-
-    public RegistryService getRegistryService() {
-        return registryService;
-    }
-}
