+++ components/bpel/pom.xml
@@ -37,7 +37,7 @@
         <module>org.wso2.carbon.bpel.cluster.notifier</module>
         <module>org.wso2.carbon.bpel.ui</module>
         <module>org.wso2.carbon.bpel.b4p</module>
-        <module>org.wso2.carbon.bpel.bam.publisher</module>
+        <module>org.wso2.carbon.bpel.analytics.publisher</module>
     </modules>
 
         <dependencies>
