+++ service-stubs/bpel/org.wso2.carbon.bpel.analytics.publisher.skeleton/src/main/resources/analytics_server_information.wsdl
@@ -1,0 +1,102 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<wsdl:definitions name="AnalyticsServerInfoManagement"
+				  targetNamespace="http://wso2.org/bpel-analytics/management/wsdl/AnalyticsServerInfoManagement"
+				  xmlns:wsdl="http://schemas.xmlsoap.org/wsdl/" 
+				  xmlns:xsd="http://www.w3.org/2001/XMLSchema" 
+				  xmlns:tns="http://wso2.org/bpel-analytics/management/wsdl/AnalyticsServerInfoManagement"
+				  xmlns:soap="http://schemas.xmlsoap.org/wsdl/soap/">
+    <wsdl:types><xsd:schema xmlns:xsd="http://www.w3.org/2001/XMLSchema" targetNamespace="http://wso2.org/bpel-analytics/management/wsdl/AnalyticsServerInfoManagement">
+	<xsd:element name="analyticsServerInformation">
+		<xsd:complexType>
+			<xsd:sequence>
+				<xsd:element name="receiverURL" type="xsd:string" minOccurs="0" maxOccurs="1"></xsd:element>
+				<xsd:element name="username" type="xsd:string" minOccurs="0" maxOccurs="1"></xsd:element>
+				<xsd:element name="password" type="xsd:string" minOccurs="0" maxOccurs="1"></xsd:element>
+			</xsd:sequence>
+		</xsd:complexType>
+	</xsd:element>
+	<xsd:element name="updateAnalyticsServerInformationResponse">
+		<xsd:complexType>
+			<xsd:sequence minOccurs="0" maxOccurs="1">
+				<xsd:element name="status" type="xsd:string"></xsd:element>
+			</xsd:sequence>
+		</xsd:complexType>
+	</xsd:element>
+	<xsd:element name="analyticsServerInformationFault">
+		<xsd:complexType>
+			<xsd:sequence>
+				<xsd:element name="analyticsServerInformationFault"
+					type="xsd:string">
+				</xsd:element>
+			</xsd:sequence>
+		</xsd:complexType>
+	</xsd:element></xsd:schema></wsdl:types>
+    <wsdl:message name="getAnalyticsServerInformationRequest">
+	</wsdl:message>
+    <wsdl:message name="getAnalyticsServerInformationResponse">
+    	<wsdl:part name="parameters" element="tns:analyticsServerInformation"></wsdl:part>
+    </wsdl:message>
+    <wsdl:message name="updateAnalyticsServerInformationResponse">
+    	<wsdl:part name="parameters" element="tns:updateAnalyticsServerInformationResponse"></wsdl:part>
+    </wsdl:message>
+    <wsdl:message name="updateAnalyticsServerInformationRequest">
+    	<wsdl:part name="parameters" element="tns:analyticsServerInformation"></wsdl:part>
+    </wsdl:message>
+    <wsdl:message name="getAnalyticsServerInformationFault">
+    	<wsdl:part name="parameters"
+    		element="tns:analyticsServerInformationFault">
+    	</wsdl:part>
+    </wsdl:message>
+    <wsdl:message name="updateAnalyticsServerInformationFault">
+    	<wsdl:part name="parameters"
+    		element="tns:analyticsServerInformationFault">
+    	</wsdl:part>
+    </wsdl:message>
+    <wsdl:portType name="AnalyticsServerInfoManagementPortType">
+    	<wsdl:operation name="getAnalyticsServerInformation">
+    		<wsdl:input message="tns:getAnalyticsServerInformationRequest"></wsdl:input>
+    		<wsdl:output message="tns:getAnalyticsServerInformationResponse"></wsdl:output>
+            <wsdl:fault name="fault" message="tns:getAnalyticsServerInformationFault"></wsdl:fault>
+        </wsdl:operation>
+    	<wsdl:operation name="updateAnalyticsServerInformation">
+    		<wsdl:input message="tns:updateAnalyticsServerInformationRequest"></wsdl:input>
+    		<wsdl:output message="tns:updateAnalyticsServerInformationResponse"></wsdl:output>
+            <wsdl:fault name="fault" message="tns:updateAnalyticsServerInformationFault"></wsdl:fault>
+        </wsdl:operation>
+    </wsdl:portType>
+    <wsdl:binding name="AnalyticsServerInfoManagementBinding"
+    	type="tns:AnalyticsServerInfoManagementPortType">
+    	<soap:binding style="document"
+    		transport="http://schemas.xmlsoap.org/soap/http" />
+    	<wsdl:operation name="getAnalyticsServerInformation">
+    		<soap:operation
+    			soapAction="http://wso2.org/bpel-analytics/management/wsdl/AnalyticsServerInfoManagement/getAnalyticsServerInformation" />
+    		<wsdl:input>
+    			<soap:body use="literal" />
+    		</wsdl:input>
+    		<wsdl:output>
+    			<soap:body use="literal" />
+    		</wsdl:output>
+    		<wsdl:fault name="fault">
+    			<soap:fault use="literal" name="fault" />
+    		</wsdl:fault>
+    	</wsdl:operation>
+    	<wsdl:operation name="updateAnalyticsServerInformation">
+    		<soap:operation
+    			soapAction="http://wso2.org/bpel-analytics/management/wsdl/AnalyticsServerInfoManagement/updateAnalyticsServerInformation" />
+    		<wsdl:input>
+    			<soap:body use="literal" />
+    		</wsdl:input>
+    		<wsdl:output>
+    			<soap:body use="literal" />
+    		</wsdl:output>
+    		<wsdl:fault name="fault">
+    			<soap:fault use="literal" name="fault" />
+    		</wsdl:fault>
+    	</wsdl:operation>
+    </wsdl:binding>
+    <wsdl:service name="AnalyticsServerInfoManagementService">
+    	<wsdl:port name="AnalyticsServerInfoManagementPort" binding="tns:AnalyticsServerInfoManagementBinding">
+    		<soap:address location="http://www.example.org/" />
+    	</wsdl:port>
+    </wsdl:service></wsdl:definitions>
