--- service-stubs/bpel/org.wso2.carbon.bpel.bam.publisher.stub/src/main/resources/bam_server_information.wsdl
@@ -1,102 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<wsdl:definitions name="BAMServerInfoManagement" 
-				  targetNamespace="http://wso2.org/bpel-bam/management/wsdl/BAMServerInfoManagement" 
-				  xmlns:wsdl="http://schemas.xmlsoap.org/wsdl/" 
-				  xmlns:xsd="http://www.w3.org/2001/XMLSchema" 
-				  xmlns:tns="http://wso2.org/bpel-bam/management/wsdl/BAMServerInfoManagement" 
-				  xmlns:soap="http://schemas.xmlsoap.org/wsdl/soap/">
-    <wsdl:types><xsd:schema xmlns:xsd="http://www.w3.org/2001/XMLSchema" targetNamespace="http://wso2.org/bpel-bam/management/wsdl/BAMServerInfoManagement">
-	<xsd:element name="bamServerInformation">
-		<xsd:complexType>
-			<xsd:sequence>
-				<xsd:element name="receiverURL" type="xsd:string" minOccurs="0" maxOccurs="1"></xsd:element>
-				<xsd:element name="username" type="xsd:string" minOccurs="0" maxOccurs="1"></xsd:element>
-				<xsd:element name="password" type="xsd:string" minOccurs="0" maxOccurs="1"></xsd:element>
-			</xsd:sequence>
-		</xsd:complexType>
-	</xsd:element>
-	<xsd:element name="updateBamServerInformationResponse">
-		<xsd:complexType>
-			<xsd:sequence minOccurs="0" maxOccurs="1">
-				<xsd:element name="status" type="xsd:string"></xsd:element>
-			</xsd:sequence>
-		</xsd:complexType>
-	</xsd:element>
-	<xsd:element name="bamServerInformationFault">
-		<xsd:complexType>
-			<xsd:sequence>
-				<xsd:element name="bamServerInformationFault"
-					type="xsd:string">
-				</xsd:element>
-			</xsd:sequence>
-		</xsd:complexType>
-	</xsd:element></xsd:schema></wsdl:types>
-    <wsdl:message name="getBamServerInformationRequest">
-	</wsdl:message>
-    <wsdl:message name="getBamServerInformationResponse">
-    	<wsdl:part name="parameters" element="tns:bamServerInformation"></wsdl:part>
-    </wsdl:message>
-    <wsdl:message name="updateBamServerInformationResponse">
-    	<wsdl:part name="parameters" element="tns:updateBamServerInformationResponse"></wsdl:part>
-    </wsdl:message>
-    <wsdl:message name="updateBamServerInformationRequest">
-    	<wsdl:part name="parameters" element="tns:bamServerInformation"></wsdl:part>
-    </wsdl:message>
-    <wsdl:message name="getBamServerInformationFault">
-    	<wsdl:part name="parameters"
-    		element="tns:bamServerInformationFault">
-    	</wsdl:part>
-    </wsdl:message>
-    <wsdl:message name="updateBamServerInformationFault">
-    	<wsdl:part name="parameters"
-    		element="tns:bamServerInformationFault">
-    	</wsdl:part>
-    </wsdl:message>
-    <wsdl:portType name="BAMServerInfoManagementPortType">
-    	<wsdl:operation name="getBamServerInformation">
-    		<wsdl:input message="tns:getBamServerInformationRequest"></wsdl:input>
-    		<wsdl:output message="tns:getBamServerInformationResponse"></wsdl:output>
-            <wsdl:fault name="fault" message="tns:getBamServerInformationFault"></wsdl:fault>
-        </wsdl:operation>
-    	<wsdl:operation name="updateBamServerInformation">
-    		<wsdl:input message="tns:updateBamServerInformationRequest"></wsdl:input>
-    		<wsdl:output message="tns:updateBamServerInformationResponse"></wsdl:output>
-            <wsdl:fault name="fault" message="tns:updateBamServerInformationFault"></wsdl:fault>
-        </wsdl:operation>
-    </wsdl:portType>
-    <wsdl:binding name="BAMServerInfoManagementBinding"
-    	type="tns:BAMServerInfoManagementPortType">
-    	<soap:binding style="document"
-    		transport="http://schemas.xmlsoap.org/soap/http" />
-    	<wsdl:operation name="getBamServerInformation">
-    		<soap:operation
-    			soapAction="http://wso2.org/bpel-bam/management/wsdl/BAMServerInfoManagement/getBamServerInformation" />
-    		<wsdl:input>
-    			<soap:body use="literal" />
-    		</wsdl:input>
-    		<wsdl:output>
-    			<soap:body use="literal" />
-    		</wsdl:output>
-    		<wsdl:fault name="fault">
-    			<soap:fault use="literal" name="fault" />
-    		</wsdl:fault>
-    	</wsdl:operation>
-    	<wsdl:operation name="updateBamServerInformation">
-    		<soap:operation
-    			soapAction="http://wso2.org/bpel-bam/management/wsdl/BAMServerInfoManagement/updateBamServerInformation" />
-    		<wsdl:input>
-    			<soap:body use="literal" />
-    		</wsdl:input>
-    		<wsdl:output>
-    			<soap:body use="literal" />
-    		</wsdl:output>
-    		<wsdl:fault name="fault">
-    			<soap:fault use="literal" name="fault" />
-    		</wsdl:fault>
-    	</wsdl:operation>
-    </wsdl:binding>
-    <wsdl:service name="BAMServerInfoManagementService">
-    	<wsdl:port name="BAMServerInfoManagementPort" binding="tns:BAMServerInfoManagementBinding">
-    		<soap:address location="http://www.example.org/" />
-    	</wsdl:port>
-    </wsdl:service></wsdl:definitions>
